Package qemu
This builds qemu via homebrew, based on upstream formula but limited to
just what we need.

Because the aarch64 macOS runners on GitHub do not support
virtualization, we hack in a few extra files to include on that
platform.  The list of files is based on what the x86_64 run had.

We need slirp because lima uses that (with `--netdev user,...`).

Signed-off-by: Mark Yen <mark.yen@suse.com>